사이드바 작업

+++ front/src/component/map/Side/Save/Content/Favorite/FDelete.js
@@ -7,19 +7,27 @@ function FDelete(props){
     const { isLoading, error, sendRequest: fetchData } = useHttp();
 
     const handleDelete = () => {
-        fetchData({url: `http://localhost:3001/store/like`, type:"delete", data:{id:props.data}}, (obj) => {
-            props.fetch()
-        })
+        if(props.type === "favorite") {
+            fetchData({url: `http://localhost:3001/store/like`, type: "delete", data: {id: props.data}}, (obj) => {
+                props.fetch()
+            })
+        }
+
+        if(props.type === "path"){
+            fetchData({url: `http://localhost:3001/store/direction`, type: "delete", data: {id: props.data}}, (obj) => {
+                props.fetch()
+            })
+        }
     }
 
     return (
         <>
             {props.data.length > 0 ?
-                    <div className={classes.fFoot}>
-                        <div onClick={handleDelete} className={classes.ffBox}>
-                            <span>저장 삭제</span>
-                        </div>
-                    </div> : ""
+                <div className={classes.fFoot}>
+                    <div onClick={handleDelete} className={classes.ffBox}>
+                        <span>저장 삭제</span>
+                    </div>
+                </div> : ""
             }
         </>
     )
